Remove AbstractPool class.
AbstractPool contains only properties at this point.
Move those properties into the connection pool impl.
Remove activate/passivate methods in favor of default no-op implementations.
